moves get_mongodb_secret to seperate file
