+++ src/client/get_deployment_id.rs
@@ -1,8 +1,7 @@
 use crate::{
     Client,
-    client::get_deployment::GetDeploymentError,
+    client::{get_deployment::GetDeploymentError, get_mongodb_secret::get_mongodb_secret},
     docker::{DockerInspectContainer, RunCommandInContainer, RunCommandInContainerError},
-    models::Deployment,
 };
 
 #[derive(Debug, thiserror::Error)]
@@ -76,35 +75,6 @@ impl<D: DockerInspectContainer + RunCommandInContainer> Client<D> {
             None => Err(GetDeploymentIdError::DeploymentIdEmpty),
         }
     }
-}
-
-pub async fn get_mongodb_secret<D: RunCommandInContainer>(
-    docker: &D,
-    deployment: &Deployment,
-    value: impl FnOnce(&Deployment) -> Option<&str>,
-    file: impl FnOnce(&Deployment) -> Option<&str>,
-) -> Result<Option<String>, RunCommandInContainerError> {
-    // Try to get the value from the environment variables first
-    if let Some(env_value) = value(deployment) {
-        return Ok(Some(env_value.to_string()));
-    }
-
-    // If the value is not found in the environment variables, try to get it from the file
-    if let Some(file_value) = file(deployment) {
-        let command_output = docker
-            .run_command_in_container(
-                &deployment.container_id,
-                vec!["cat".to_string(), file_value.to_string()],
-            )
-            .await?;
-
-        if let Some(line) = command_output.stdout.into_iter().next() {
-            return Ok(Some(line));
-        }
-    }
-
-    // If the value is not found in the environment variables or the file, return None
-    Ok(None)
 }
 
 #[cfg(test)]
